change shadowmap to R32, support hw filter

+++ src/Assets/CPUAccelerationStructure.h
@@ -94,7 +94,7 @@ private:
 
     std::vector<std::tuple<glm::ivec3, ECubeProcType> > needUpdateGroups;
 
-    std::vector<uint32_t> shadowMapR32;
+    std::vector<float> shadowMapR32;
     bool generatingShadowMap = false;
 
     bool needFlush = false;
